+++ src/main/java/hana/HanaException.java
@@ -1,3 +1,5 @@
+package hana;
+
 public class HanaException extends Exception{
     public HanaException(String message) {
         super(message);
