+++ src/main/java/hana/command/DeadlineCommand.java
@@ -1,3 +1,12 @@
+package hana.command;
+
+import hana.tasklist.TaskList;
+import hana.task.Task;
+import hana.task.Deadline;
+import hana.ui.Ui;
+import hana.storage.Storage;
+import hana.HanaException;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
