+++ src/main/java/hana/command/EventCommand.java
@@ -1,3 +1,12 @@
+package hana.command;
+
+import hana.HanaException;
+import hana.storage.Storage;
+import hana.task.Event;
+import hana.task.Task;
+import hana.tasklist.TaskList;
+import hana.ui.Ui;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
