+++ src/main/java/hana/command/helpCommand.java
@@ -1,3 +1,10 @@
+package hana.command;
+
+import hana.HanaException;
+import hana.storage.Storage;
+import hana.tasklist.TaskList;
+import hana.ui.Ui;
+
 public class helpCommand extends Command {
 
 	@Override
