+++ src/main/java/hana/task/Event.java
@@ -1,3 +1,5 @@
+package hana.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
