+++ src/main/java/hana/task/Task.java
@@ -1,4 +1,6 @@
-abstract class Task {
+package hana.task;
+
+public abstract class Task {
     protected String description;
     protected boolean isDone;
